修改删除资料的漏洞

+++ src/AppBundle/Controller/Course/CourseSetFileManageController.php
@@ -159,12 +159,15 @@ class CourseSetFileManageController extends BaseController
         if ($request->getMethod() == 'POST') {
             $formData = $request->request->all();
 
-            $this->getMaterialService()->deleteMaterials($id, $formData['ids']);
+            $deletedMaterials = $this->getMaterialService()->deleteMaterials($id, $formData['ids']);
 
-            if (isset($formData['isDeleteFile']) && $formData['isDeleteFile']) {
-                foreach ($formData['ids'] as $key => $fileId) {
-                    if ($this->getUploadFileService()->canManageFile($fileId)) {
-                        $this->getUploadFileService()->deleteFile($fileId);
+            if (!empty($deletedMaterials)) {
+                $fileIds = array_unique(ArrayToolkit::column($deletedMaterials, 'fileId'));
+                if (isset($formData['isDeleteFile']) && $formData['isDeleteFile']) {
+                    foreach ($fileIds as $key => $fileId) {
+                        if ($this->getUploadFileService()->canManageFile($fileId)) {
+                            $this->getUploadFileService()->deleteFile($fileId);
+                        }
                     }
                 }
             }
